+++ php/enqueue-leafletplugins.php
@@ -47,7 +47,7 @@ function leafext_enqueue_elevation() {
 			'leaflet-plugins/Leaflet.AlmostOver/leaflet.almostover.js',
 			LEAFEXT_PLUGIN_FILE
 		),
-		array( 'Leaflet_GeometryUtil' ),
+		array( 'Leaflet_GeometryUtil', 'wp_leaflet_map' ),
 		null,
 		true
 	);
@@ -81,7 +81,7 @@ function leafext_enqueue_elevation() {
 		null
 	);
 	leafext_enqueue_css();
-	if ( get_locale() === 'sv_SE' || get_locale() === 'it_IT' || get_locale() === 'ca' ) {
+	if ( file_exists( LEAFEXT_PLUGIN_DIR . '/lang/extensions-leaflet-map-' . get_locale() . '-elevation_js.json' ) ) {
 		wp_set_script_translations( 'elevation_js', 'extensions-leaflet-map', LEAFEXT_PLUGIN_DIR . '/lang/' );
 	}
 }