+++ CHANGELOG.md
@@ -1,8 +1,10 @@
-### 4.3.2 / 2407xx
+### 4.4 / 240804
 
 * leaflet-parentgroup: you can set more options now.
 * leaflet-extramarker: title with spaces fixed.
-* overviewmap: marker title
+* overviewmap: marker title, categories option AND
+* targetmarker, targetlink: new option - mapid
+* elevation: new languages: portugese, polish
 
 ### 4.3.1 / 240701