+++ changes.md
@@ -3,6 +3,7 @@
 * leaflet-parentgroup: you can set more options now.
 * leaflet-extramarker: title with spaces fixed.
 * overviewmap: marker title
+* targetmarker, targetlink: new option - mapid
 
 Please note the [install instructions](https://leafext.de/en/doku/about/versions/).